requirements.txt
Django==5.1.4
pydantic==2.10.3
pydantic_settings==2.6.1
psycopg2_binary==2.9.10
gunicorn==23.0.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.3.1
redis==5.2.1
pytz==2022.7